Extract ACP initialize frame helper
Define the initialize-frame skeleton once and pass only the capability payload
into tests that vary ACP client capabilities. Move forwarding-specific ACP tests
into a child module so the main ACP test file stays below the repository line
budget while preserving the same private test helpers.
